+++ src/modules/adempiere-api/api/extensions/adempiere/dashboard/addons/user.js
@@ -1,12 +1,29 @@
+/************************************************************************************
+ * Copyright (C) 2012-2022 E.R.P. Consultores y Asociados, C.A.                     *
+ * Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com                     *
+ * This program is free software: you can redistribute it and/or modify             *
+ * it under the terms of the GNU General Public License as published by             *
+ * the Free Software Foundation, either version 2 of the License, or                *
+ * (at your option) any later version.                                              *
+ * This program is distributed in the hope that it will be useful,                  *
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of                   *
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                     *
+ * GNU General Public License for more details.                                     *
+ * You should have received a copy of the GNU General Public License                *
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.            *
+ ************************************************************************************/
+
 import { Router } from 'express';
 import {
-  convertFavoriteFromGRPC,
+  convertFavoriteFromGRPC
+} from '@adempiere/grpc-api/src/utils/dashboardingFromGRPC';
+import {
   convertRecentItemFromGRPC
-} from '@adempiere/grpc-api/lib/convertBusinessData';
+} from '@adempiere/grpc-api/src/utils/logsFromGRPC';
 
 module.exports = ({ config }) => {
   const api = Router();
-  const ServiceApi = require('@adempiere/grpc-api')
+  const ServiceApi = require('@adempiere/grpc-api/src/services/dashboarding')
   const service = new ServiceApi(config)
 
   /**
